Move some AR test cases to inheritance_test.rb
These methods are defined in inheritance.rb

* `abstract_class?`
* `descends_from_active_record?`
* `compute_type`
